requirements.txt
aiogram~=3.17.0
asyncpg~=0.30.0
loguru~=0.7.3
pyyaml~=6.0.2
pydantic~=2.10.6
python-dotenv~=1.0.1
redis~=5.2.1